refactor venue lookup dictionaries to public class
